requirements.txt
cerebras-cloud-sdk==1.3.0
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==20.1.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
requests==2.32.3
scipy==1.13.1
Werkzeug==3.0.4
zipp==3.20.2
